refactor(vm): move the host↔agent wire vocabulary into arcbox-vm-proto
The boot parameters, exec-channel frames, and file-channel frames that
the sandbox manager and the in-sandbox vm-agent must agree on now live
in a serde-only crate both link. arcbox-vm re-exports them as
`boot_proto`, `file_io::proto`, and the `vsock` constants/DTOs so every
existing import keeps resolving; the vm-agent binary still builds
unchanged from arcbox-vm at this step and moves out next.

The one shared `StartCommand` carries the agent's decoder defaults, so
the tolerant decode a restored snapshot's old agent relies on is defined
once instead of mirrored by hand.

Refs CORE-127.
