+++ virt/arcbox-vm-proto/src/boot.rs
@@ -12,11 +12,11 @@ use std::str::FromStr;
 ///
 /// Format: `ip=<client>::<gateway>:<netmask>::eth0:off`
 ///
-/// Constructed by [`SandboxManager`](crate::SandboxManager) when building
-/// boot args, and parsed by `vm-agent` to derive the DNS nameserver
-/// from the gateway.
+/// Constructed by the sandbox manager (`arcbox_vm::SandboxManager`) when
+/// building boot args, and parsed by `vm-agent` to derive the DNS
+/// nameserver from the gateway.
 ///
-/// [`Display`] and [`FromStr`] round-trip through the same format so the
+/// [`fmt::Display`] and [`FromStr`] round-trip through the same format so the
 /// encoding is defined exactly once.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct KernelIpParam {
